backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai>=1.30.0
python-dotenv